Sample PCA plot for transformed data

Plots the results of PCA on a 3-dimensional space, interactively

pcaplot3d(x, intgroup = "condition", ntop = 500, returnData = FALSE,
  title = NULL, pcX = 1, pcY = 2, pcZ = 3, text_labels = TRUE,
  point_size = 3)

Arguments

x
A DESeqTransform object, with data in assay(x), produced for example by either rlog or varianceStabilizingTransformation

intgroup
Interesting groups: a character vector of names in colData(x) to use for grouping

ntop
Number of top genes to use for principal components, selected by highest row variance

returnData
logical, if TRUE returns a data.frame for further use, containing the selected principal components and intgroup covariates for custom plotting

title
The plot title

pcX
The principal component to display on the x axis

pcY
The principal component to display on the y axis

pcZ
The principal component to display on the z axis

text_labels
Logical, whether to display the labels with the sample identifiers

point_size
Integer, the size of the points for the samples

Value

A html-based visualization of the 3d PCA plot

Examples

dds <- makeExampleDESeqDataSet_multifac(betaSD_condition = 3,betaSD_tissue = 1)
rlt <- DESeq2::rlogTransformation(dds)
pcaplot3d(rlt, ntop=200)